requirements.txt
streamlit==1.39.0
pandas==2.3.3
dnspython==2.7.1
email-validator==2.2.1
ipwhois==1.3.1
requests==2.32.5
rich==14.1.1
python-dotenv==1.1.1
pydantic==2.11.7
plotly==6.2.0
pytest==8.4.1
pip-tools==7.4.1
